requirements.txt

fastapi
uvicorn
pydantic
faiss-cpu
numpy
sentence-transformers
python-dotenv
markdownify
beautifulsoup4
playwright
